+++ packages/runtime-core/src/helpers/withMemo.ts
@@ -7,28 +7,30 @@ export function withMemo(
   index: number
 ) {
   const cached = cache[index] as VNode | undefined
-  if (cached && isMemoSame(cached.memo!, memo)) {
-    // make sure to let parent block track it when returning cached
-    if (isBlockTreeEnabled > 0 && currentBlock) {
-      currentBlock.push(cached)
-    }
+  if (cached && isMemoSame(cached, memo)) {
     return cached
   }
   const ret = render()
 
   // shallow clone
-  ret.memo = memo.slice(0, memo.length)
+  ret.memo = memo.slice()
   return (cache[index] = ret)
 }
 
-export function isMemoSame(prev: any[], next: any[]) {
+export function isMemoSame(cached: VNode, memo: any[]) {
+  const prev: any[] = cached.memo!
   if (prev.length != next.length) {
     return false
   }
   for (let i = 0; i < prev.length; i++) {
-    if (prev[i] !== next[i]) {
+    if (prev[i] !== memo[i]) {
       return false
     }
+  }
+
+  // make sure to let parent block track it when returning cached
+  if (isBlockTreeEnabled > 0 && currentBlock) {
+    currentBlock.push(cached)
   }
   return true
 }